Hierarchy

Functions for navigating the parent/child cell hierarchy and compact/uncompact operations.

Corresponds to the H3 C API Hierarchy section.

Parent / Child Navigation

H3X.cellToParent — Function

cellToParent(h::H3Index, parentRes::Int) -> Tuple{H3Error, H3Index}

Get the parent cell at the given coarser resolution.

See also the H3 C API: cellToParent

H3X.cellToCenterChild — Function

cellToCenterChild(h::H3Index, childRes::Int) -> Tuple{H3Error, H3Index}

Get the center child cell at the given finer resolution.

See also the H3 C API: cellToCenterChild

H3X.cellToChildren — Function

cellToChildren(h::H3Index, childRes::Int) -> Tuple{H3Error, Vector{H3Index}}

Get all children of a cell at a given finer resolution.

See also the H3 C API: cellToChildren

H3X.cellToChildrenSize — Function

cellToChildrenSize(h::H3Index, childRes::Int) -> Tuple{H3Error, Int64}

Get the number of children at the given finer resolution.

See also the H3 C API: cellToChildrenSize

H3X.cellToChildPos — Function

cellToChildPos(child::H3Index, parentRes::Int) -> Tuple{H3Error, Int64}

Get the position of a child cell within its parent's ordered children.

See also the H3 C API: cellToChildPos

H3X.childPosToCell — Function

childPosToCell(childPos::Int64, parent::H3Index, childRes::Int) -> Tuple{H3Error, H3Index}

Convert a child position to the corresponding child cell.

See also the H3 C API: childPosToCell

Compact / Uncompact

H3X.compactCells — Function

compactCells(h3Set::Vector{H3Index}) -> Tuple{H3Error, Vector{H3Index}}

Compact a set of cells into the smallest equivalent set.

See also the H3 C API: compactCells

H3X.uncompactCells — Function

uncompactCells(compactedSet::Vector{H3Index}, res::Int) -> Tuple{H3Error, Vector{H3Index}}

Uncompact a set of cells to the given resolution.

See also the H3 C API: uncompactCells

H3X.uncompactCellsSize — Function

uncompactCellsSize(compactedSet::Vector{H3Index}, res::Int) -> Tuple{H3Error, Int64}

Get the number of cells that would result from uncompacting to the given resolution.

See also the H3 C API: uncompactCellsSize

Resolution 0 and Pentagons

H3X.getRes0Cells — Function

getRes0Cells() -> Tuple{H3Error, Vector{H3Index}}

Get all resolution 0 (base cell) indexes.

See also the H3 C API: getRes0Cells

H3X.getPentagons — Function

getPentagons(res::Int) -> Tuple{H3Error, Vector{H3Index}}

Get all pentagon cell indexes at the given resolution.

See also the H3 C API: getPentagons

H3X.pentagonCount — Function

pentagonCount() -> Int

Return the number of pentagon cells per resolution (always 12).

See also the H3 C API: pentagonCount